Refactor database migration logic
